Move SoundSpeedSquared compute tag to ComputeItems file
