+++ src/Evolution/Systems/NewtonianEuler/Limiters/CharacteristicHelpers.cpp
@@ -20,7 +20,7 @@
 #include "Domain/Tags.hpp"               // IWYU pragma: keep
 #include "ErrorHandling/Assert.hpp"
 #include "Evolution/Systems/NewtonianEuler/Characteristics.hpp"
-#include "Evolution/Systems/NewtonianEuler/SoundSpeedSquared.hpp"
+#include "Evolution/Systems/NewtonianEuler/ComputeItems.hpp"
 #include "Evolution/Systems/NewtonianEuler/Tags.hpp"
 #include "NumericalAlgorithms/Spectral/Mesh.hpp"
 #include "Utilities/GenerateInstantiations.hpp"
