+++ tests/Unit/Evolution/Systems/NewtonianEuler/CMakeLists.txt
@@ -9,7 +9,6 @@ set(LIBRARY_SOURCES
   Test_ConservativeFromPrimitive.cpp
   Test_Fluxes.cpp
   Test_PrimitiveFromConservative.cpp
-  Test_SoundSpeedSquared.cpp
   Test_Sources.cpp
   Test_Tags.cpp
   Test_TimeDerivativeTerms.cpp
